move parseSnapshotInputs
