+++ packages/sdk/src/evm/common/snapshots.ts
@@ -10,21 +10,7 @@ import {
 } from "./sharded-merkle-tree";
 import { ThirdwebStorage } from "@thirdweb-dev/storage";
 import { ethers } from "ethers";
-
-export async function parseSnapshotInputs(inputs: SnapshotInput) {
-  const chunkSize = 25000;
-  const chunks = Array.from(
-    { length: Math.ceil(inputs.length / chunkSize) },
-    (_, i) => inputs.slice(i * chunkSize, i * chunkSize + chunkSize),
-  );
-
-  const results = [];
-  for (const chunk of chunks) {
-    results.push(...(await SnapshotInputSchema.parseAsync(chunk)));
-  }
-
-  return results;
-}
+import { parseSnapshotInputs } from "./parseSnapshotInputs";
 
 /**
  * Create a snapshot (merkle tree) from a list of addresses and uploads it to IPFS
